Separate mime bundle rendering in xmime.hpp
